+++ src/birth/game-play-initializer.cpp
@@ -6,7 +6,7 @@
 #include "info-reader/fixed-map-parser.h"
 #include "inventory/inventory-slot-types.h"
 #include "market/arena-entry.h"
-#include "market/arena.h"
+#include "market/melee-arena.h"
 #include "pet/pet-util.h"
 #include "player-base/player-class.h"
 #include "player-base/player-race.h"
