+++ src/dungeon/dungeon-processor.cpp
@@ -18,7 +18,7 @@
 #include "io/cursor.h"
 #include "io/input-key-requester.h"
 #include "io/write-diary.h"
-#include "market/arena.h"
+#include "market/melee-arena.h"
 #include "mind/mind-ninja.h"
 #include "monster/monster-compaction.h"
 #include "monster/monster-processor.h"
