+++ src/market/arena.h
@@ -2,5 +2,3 @@
 
 class PlayerType;
 bool arena_comm(PlayerType *player_ptr, int cmd);
-void update_melee_gladiators(PlayerType *player_ptr);
-bool monster_arena_comm(PlayerType *player_ptr);
