+++ src/market/melee-arena.h
@@ -1,0 +1,5 @@
+#pragma once
+
+class PlayerType;
+void update_melee_gladiators(PlayerType *player_ptr);
+bool monster_arena_comm(PlayerType *player_ptr);
