+++ src/world/world-turn-processor.cpp
@@ -15,8 +15,8 @@
 #include "inventory/inventory-curse.h"
 #include "inventory/recharge-processor.h"
 #include "io/write-diary.h"
-#include "market/arena.h"
 #include "market/bounty.h"
+#include "market/melee-arena.h"
 #include "monster-floor/monster-generator.h"
 #include "monster-floor/monster-summon.h"
 #include "monster/monster-describer.h"
